+++ Config.lua
@@ -102,16 +102,42 @@ config:SetScript("OnShow", function(self)
 		self.tabButtonSettings:Enable()
 	end)
 
-	-- DESCRIPTION
-	local hexColor = toHex(self.config.lineColor)
-	local description = self.generalPanel:CreateFontString(nil, "ARTWORK", "GameFontHighlight")
-	description:SetPoint("TOPLEFT", 8, -10)
-	description:SetJustifyH("LEFT")
-	description:SetText(L["SETTINGS_DESCRIPTION"]:format(hexColor))
+	-- EXPAND TO TEXT
+	local expandToText = self.generalPanel:CreateFontString(nil, "ARTWORK", "GameFontHighlight")
+	expandToText:SetPoint("TOPRIGHT", -10, -8)
+	expandToText:SetWidth(114)
+	expandToText:SetText(L["Expand to"])
+
+	-- EXPAND TO COMBOBOX
+	local expandToCombobox = CreateFrame("FRAME", "HidingBarAddonExpandTo", self.generalPanel, "UIDropDownMenuTemplate")
+	expandToCombobox:SetPoint("TOPRIGHT", expandToText, "BOTTOMRIGHT", 17, -2)
+	UIDropDownMenu_SetWidth(expandToCombobox, 100)
+
+	local function updateExpandTo(btn)
+		UIDropDownMenu_SetSelectedValue(expandToCombobox, btn.value)
+		self.hidingBar:setBarExpand(btn.value)
+	end
+
+	UIDropDownMenu_Initialize(expandToCombobox, function(self, level)
+		local info = UIDropDownMenu_CreateInfo()
+
+		info.checked = nil
+		info.text = L["Right / Bottom"]
+		info.value = 0
+		info.func = updateExpandTo
+		UIDropDownMenu_AddButton(info)
+
+		info.checked = nil
+		info.text = L["Left / Top"]
+		info.value = 1
+		info.func = updateExpandTo
+		UIDropDownMenu_AddButton(info)
+	end)
+	UIDropDownMenu_SetSelectedValue(expandToCombobox, self.config.expand)
 
 	-- LINE COLOR
 	local lineColor = CreateFrame("BUTTON", nil, self.generalPanel, "HidingBarAddonColorButton")
-	lineColor:SetPoint("TOPRIGHT", -8, -8)
+	lineColor:SetPoint("TOPRIGHT", expandToCombobox, "BOTTOMRIGHT", -18, -2)
 	lineColor.color:SetColorTexture(unpack(self.config.lineColor))
 	local lineColorText = self.generalPanel:CreateFontString(nil, "ARTWORK", "GameFontHighlight")
 	lineColorText:SetPoint("RIGHT", lineColor, "LEFT", -3, 0)
@@ -122,7 +148,7 @@ config:SetScript("OnShow", function(self)
 		self.config.lineColor = {ColorPickerFrame:GetColorRGB()}
 		self.hidingBar.drag.bg:SetColorTexture(ColorPickerFrame:GetColorRGB())
 		local hexColor = toHex(self.config.lineColor)
-		description:SetText(L["SETTINGS_DESCRIPTION"]:format(hexColor))
+		self.description:SetText(L["SETTINGS_DESCRIPTION"]:format(hexColor))
 		self.fade.Text:SetText(L["Fade out line"]:format(hexColor))
 		self.lineWidth.text:SetText(L["Line width"]:format(hexColor))
 		lineColor.color:SetColorTexture(ColorPickerFrame:GetColorRGB())
@@ -133,7 +159,7 @@ config:SetScript("OnShow", function(self)
 		self.config.lineColor = {color.r, color.g, color.b}
 		self.hidingBar.drag.bg:SetColorTexture(color.r, color.g, color.b)
 		local hexColor = toHex(self.config.lineColor)
-		description:SetText(L["SETTINGS_DESCRIPTION"]:format(hexColor))
+		self.description:SetText(L["SETTINGS_DESCRIPTION"]:format(hexColor))
 		self.fade.Text:SetText(L["Fade out line"]:format(hexColor))
 		self.lineWidth.text:SetText(L["Line width"]:format(hexColor))
 		lineColor.color:SetColorTexture(color.r, color.g, color.b)
@@ -195,9 +221,16 @@ config:SetScript("OnShow", function(self)
 		OpenColorPicker(btn)
 	end)
 
+	-- DESCRIPTION
+	local hexColor = toHex(self.config.lineColor)
+	self.description = self.generalPanel:CreateFontString(nil, "ARTWORK", "GameFontHighlight")
+	self.description:SetPoint("TOPLEFT", 8, -10)
+	self.description:SetJustifyH("LEFT")
+	self.description:SetText(L["SETTINGS_DESCRIPTION"]:format(hexColor))
+
 		-- ORIENTATION TEXT
 	local orientationText = self.generalPanel:CreateFontString(nil, "ARTWORK", "GameFontHighlight")
-	orientationText:SetPoint("TOPLEFT", description, "BOTTOMLEFT", 0, -23)
+	orientationText:SetPoint("TOPLEFT", self.description, "BOTTOMLEFT", 0, -23)
 	orientationText:SetText(L["Orientation"])
 
 	-- ORIENTATION COMBOBOX
@@ -237,7 +270,7 @@ config:SetScript("OnShow", function(self)
 	-- FRAME STARTA TEXT
 	local fsText = self.generalPanel:CreateFontString(nil, "ARTWORK", "GameFontHighlight")
 	fsText:SetPoint("LEFT", orientationCombobox, "RIGHT", -5, 0)
-	fsText:SetText(L["Panel level"])
+	fsText:SetText(L["Strata of panel"])
 
 	-- FRAME STRATA COMBOBOX
 	local fsCombobox =  CreateFrame("FRAME", "HidingBarAddonFrameStrata", self.generalPanel, "UIDropDownMenuTemplate")
@@ -291,7 +324,6 @@ config:SetScript("OnShow", function(self)
 			UIFrameFadeRemoveFrame(self.hidingBar.drag)
 			self.hidingBar.drag:SetAlpha(1)
 		end
-		self.hidingBar:applyLayout()
 	end)
 
 	-- FADE OPACITY
@@ -330,7 +362,7 @@ config:SetScript("OnShow", function(self)
 
 	-- SHOW HANDLER TEXT
 	local showHandlerText = self.generalPanel:CreateFontString(nil, "ARTWORK", "GameFontHighlight")
-	showHandlerText:SetPoint("TOPLEFT", self.lineWidth, "BOTTOMLEFT", 0, -18)
+	showHandlerText:SetPoint("TOPLEFT", self.lineWidth, "BOTTOMLEFT", 0, -20)
 	showHandlerText:SetText(L["Show on"])
 
 	-- SHOW HANDLER
@@ -513,14 +545,14 @@ config:SetScript("OnShow", function(self)
 		if self.config.size ~= value then
 			self.config.size = value
 			slider.label:SetText(value)
-			self:applyLayout()
+			self:applyLayout(.3)
 			self:hidingBarUpdate()
 		end
 	end)
 
 	-- SLIDER BUTTONS SIZE
 	local buttonSize = CreateFrame("SLIDER", nil, self.buttonSettingsPanel, "HidingBarAddonSliderTemplate")
-	buttonSize:SetPoint("TOPLEFT", buttonNumber, "BOTTOMLEFT", 0, -15)
+	buttonSize:SetPoint("TOPLEFT", buttonNumber, "BOTTOMLEFT", 0, -18)
 	buttonSize:SetPoint("RIGHT", -30, 0)
 	buttonSize:SetPoint("RIGHT", -30, 0)
 	buttonSize:SetMinMaxValues(16, 64)
@@ -534,11 +566,45 @@ config:SetScript("OnShow", function(self)
 			self.config.buttonSize = value
 			slider.label:SetText(value)
 			self:setButtonSize()
-			self:applyLayout()
+			self:applyLayout(.3)
 			self.hidingBar:setButtonSize()
 			self:hidingBarUpdate()
 		end
 	end)
+
+	-- POSTION OF MINIMAP BUTTON TEXT
+	local mbtnPostionText = self.buttonSettingsPanel:CreateFontString(nil, "ARTWORK", "GameFontHighlight")
+	mbtnPostionText:SetPoint("TOPLEFT", buttonSize, "BOTTOMLEFT", 0, -20)
+	mbtnPostionText:SetText(L["Position of minimap buttons"])
+
+	-- POSITION OF MINIMAP BUTTON
+	local mbtnPostionCombobox = CreateFrame("FRAME", "HidingBarAddonMBtnPosition", self.buttonSettingsPanel, "UIDropDownMenuTemplate")
+	mbtnPostionCombobox:SetPoint("LEFT", mbtnPostionText, "RIGHT", -12, 0)
+	UIDropDownMenu_SetWidth(mbtnPostionCombobox, 100)
+
+	local function updateMBtnPostion(btn)
+		UIDropDownMenu_SetSelectedValue(mbtnPostionCombobox, btn.value)
+		self.config.mbtnPosition = btn.value
+		self:hidingBarUpdate()
+		self:applyLayout(.3)
+	end
+
+	UIDropDownMenu_Initialize(mbtnPostionCombobox, function(self, level)
+		local info = UIDropDownMenu_CreateInfo()
+
+		info.checked = nil
+		info.text = L["A new line"]
+		info.value = 0
+		info.func = updateMBtnPostion
+		UIDropDownMenu_AddButton(info)
+
+		info.checked = nil
+		info.text = L["Followed"]
+		info.value = 1
+		info.func = updateMBtnPostion
+		UIDropDownMenu_AddButton(info)
+	end)
+	UIDropDownMenu_SetSelectedValue(mbtnPostionCombobox, self.config.mbtnPosition)
 
 	-- BUTTONS TAB PANEL
 	self.buttonsTabPanel = CreateFrame("FRAME", nil, self, "HidingBarAddonOptionsPanel")
@@ -693,6 +759,7 @@ function config:addIgnoreName(name)
 	tinsert(self.config.ignoreMBtn, name)
 	sort(self.config.ignoreMBtn)
 	self.ignoreScroll:update()
+	StaticPopup_Show(config.addonName.."GET_RELOAD")
 end
 
 
@@ -721,38 +788,40 @@ end
 function config:dragBtn(btn)
 	local scale = btn:GetScale()
 	local x = btn:GetLeft() - (self.buttonPanel:GetLeft() + 4) / scale
-	local y = (self.buttonPanel:GetTop() - 4 - btn.offset) / scale - btn:GetTop()
+	local y = (self.buttonPanel:GetTop() - 4) / scale - btn:GetTop()
 	local buttonSize = self.config.buttonSize / scale
 	local row, column = math.floor(y / buttonSize + .5), math.floor(x / buttonSize + .5) + 1
-	if row < 0 then row = 0 end
-	if row > btn.maxRow then row = btn.maxRow end
-	if column < 1 then column = 1 end
-	if column > btn.maxColumn then column = btn.maxColumn end
-	local order = row * self.size + column
-	if order > #btn.btnList then order = #btn.btnList end
+	if row < btn.minRow then row = btn.minRow
+	elseif row > btn.maxRow then row = btn.maxRow end
+	if column < 1 then column = 1
+	elseif column > btn.maxColumn then column = btn.maxColumn end
+	local order = row * self.size + column - btn.orderDelta
+	if order < 1 then order = 1
+	elseif order > #btn.btnList then order = #btn.btnList end
 
 	local step = order > btn.settings[2] and 1 or -1
 	for i = btn.settings[2], order - step, step do
 		local button = btn.btnList[i + step]
 		btn.btnList[i] = button
 		button.settings[2] = i
-		self:setPointBtn(button, 4, 4 + btn.offset, i, .1)
+		self:setPointBtn(button, 4, 4, i + btn.orderDelta, .1)
 	end
 	btn.btnList[order] = btn
 	btn.settings[2] = order
 end
 
 
-function config:dragStart(btn, offset)
+function config:dragStart(btn, orderDelta)
 	btn.isDrag = true
 	for i = 1, #btn.btnList do
 		btn.btnList[i].settings[2] = i
 	end
 	if not btn.level then btn.level = btn:GetFrameLevel() end
 	btn:SetFrameLevel(btn.level + 2)
-	btn.offset = offset or 0
-	btn.maxColumn = #btn.btnList
-	btn.maxRow = math.floor((btn.maxColumn - 1) / self.size)
+	btn.orderDelta = orderDelta or 0
+	btn.maxColumn = #btn.btnList + btn.orderDelta
+	btn.minRow = math.floor(btn.orderDelta / self.size)
+	btn.maxRow = math.ceil(btn.maxColumn / self.size) - 1
 	if btn.maxColumn > self.size then btn.maxColumn = self.size end
 	btn:SetScript("OnUpdate", function(btn) self:dragBtn(btn) end)
 	btn:StartMoving()
@@ -764,7 +833,7 @@ function config:dragStop(btn)
 	btn:SetFrameLevel(btn.level)
 	btn:SetScript("OnUpdate", nil)
 	btn:StopMovingOrSizing()
-	self:setPointBtn(btn, 4, 4 + btn.offset, btn.settings[2], .3)
+	self:setPointBtn(btn, 4, 4, btn.settings[2] + btn.orderDelta, .3)
 	self.hidingBar:sort()
 	self:hidingBarUpdate()
 end
@@ -788,6 +857,14 @@ do
 		config:hidingBarUpdate()
 	end
 
+	local function btnDragStart(btn)
+		config:dragStart(btn)
+	end
+
+	local function btnDragStop(btn)
+		config:dragStop(btn)
+	end
+
 	function config:createButton(name, button, update)
 		if not self.buttonPanel or buttonsByName[name] then return end
 		local btn = CreateFrame("CheckButton", nil, self.buttonPanel, "HidingBarAddonConfigButtonTemplate")
@@ -803,8 +880,8 @@ do
 		btn.iconDesaturated = button.iconDesaturated
 		btn.btnList = self.buttons
 		btn:HookScript("OnClick", btnClick)
-		btn:SetScript("OnDragStart", function(btn) self:dragStart(btn) end)
-		btn:SetScript("OnDragStop", function(btn) self:dragStop(btn) end)
+		btn:SetScript("OnDragStart", btnDragStart)
+		btn:SetScript("OnDragStop", btnDragStop)
 		btn.settings = self.config.btnSettings[name]
 		btn:SetChecked(btn.settings[1])
 		buttonsByName[name] = btn
@@ -828,9 +905,16 @@ do
 			self:SetChecked(not self:GetChecked())
 			StaticPopup_Show(config.addonName.."ADD_IGNORE_MBTN", NORMAL_FONT_COLOR_CODE..self.name..FONT_COLOR_CODE_CLOSE, nil, function()
 				config:addIgnoreName(self.name)
-				StaticPopup_Show(config.addonName.."GET_RELOAD")
 			end)
 		end
+	end
+
+	local function btnDragStart(btn)
+		config:dragStart(btn, config.orderMBtnDelta)
+	end
+
+	local function btnDragStop(btn)
+		config:dragStop(btn)
 	end
 
 	function config:createMButton(name, icon)
@@ -843,8 +927,8 @@ do
 		btn.color = {icon:GetVertexColor()}
 		btn.btnList = self.mbuttons
 		btn:HookScript("OnClick", btnClick)
-		btn:SetScript("OnDragStart", function(btn) self:dragStart(btn, math.ceil(#self.buttons / self.size) * self.config.buttonSize) end)
-		btn:SetScript("OnDragStop", function(btn) self:dragStop(btn) end)
+		btn:SetScript("OnDragStart", btnDragStart)
+		btn:SetScript("OnDragStop", btnDragStop)
 		btn.settings = self.config.mbtnSettings[name]
 		btn:SetChecked(btn.settings[1])
 		buttonsByName[name] = btn
@@ -925,22 +1009,24 @@ function config:setPointBtn(btn, offsetX, offsetY, order, delay)
 end
 
 
-function config:applyLayout()
+function config:applyLayout(delay)
+	local offsetX, offsetY = 4, 4
 	local maxColumns = math.floor(560 / self.config.buttonSize)
 	self.size = self.config.size
 	if self.size > maxColumns then self.size = maxColumns end
 
 	for i, btn in ipairs(self.buttons) do
-		self:setPointBtn(btn, 4, 4, i)
+		self:setPointBtn(btn, offsetX, offsetY, i, delay)
 	end
-	local rows = math.ceil(#self.buttons / self.size)
-	local offsetY = rows * self.config.buttonSize + 4
+	self.orderMBtnDelta = self.config.mbtnPosition == 1 and #self.buttons or math.ceil(#self.buttons / self.size) * self.size
 	for i, btn in ipairs(self.mbuttons) do
-		self:setPointBtn(btn, 4, offsetY, i)
+		self:setPointBtn(btn, offsetX, offsetY, i + self.orderMBtnDelta, delay)
 	end
 
-	rows = rows + math.ceil(#self.mbuttons / self.size)
-	self.buttonPanel:SetSize(self.size * self.config.buttonSize + 8, rows * self.config.buttonSize + 8)
+	local rows = math.ceil((#self.mbuttons + self.orderMBtnDelta) / self.size)
+	local width = self.size * self.config.buttonSize + offsetX * 2
+	local height = rows * self.config.buttonSize +  offsetY * 2
+	self.buttonPanel:SetSize(width, height)
 end