+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-* Added "Queue status" to grab default buttons.
-* Added "Panel level" option.
-* Fixed "Couldn't find inherited node: BackdropTemplate".
+* Added option "Position of minimap buttons".
+* Added option "Expand to".